refactor(client): components を lobby / auction でドメイン分割
- game/ を auction/ に統合(意味重複の解消)
- ゲーム開始前画面(Home/Lobby/JoinForm/NameRegister)を lobby/ に集約
- ゲーム中画面(AuctionBoard・EndedScreen 含む)を auction/ に集約
- GameBoard を AuctionBoard にリネーム
- sketch/ や rules/、共通の RulesLauncher/RulesLink/UserBadge は据え置き
