+++ packages/client/src/components/auction/AuctionCardArt.tsx
