+++ packages/client/src/components/auction/CurrentBidDisplay.tsx
