+++ packages/client/src/components/auction/DeclarationBanner.tsx
