+++ packages/client/src/components/auction/MyStatsBar.tsx
