+++ packages/client/src/components/auction/OpponentStrip.tsx
